requirements.txt
pytest>=7.0.0
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest-html>=4.1.0